+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Explicitly delete `ProjectVPC` by `ID` to avoid conflicts 
 - Speed up `ProjectVPC` deletion by exiting on `DELETING` status
 - Fix missing RBAC permissions to update finalizers for various controllers 
+- Refactor `ClickhouseUser` controller
+- Mark `ClickhouseUser.spec.project` and `ClickhouseUser.spec.serviceName` as immutable
 
 ## v0.9.0 - 2023-03-03